add: base class for entities
