+++ app/Entities/CalendarDayEntity.php
@@ -11,7 +11,7 @@ use Miser\Exceptions\InvalidEntityParameterException;
 /**
  * Class CalendarDayEntity
  */
-class CalendarDayEntity
+class CalendarDayEntity extends Entity
 {
     /** @var int  Number of the day. */
     protected $day;
@@ -40,33 +40,5 @@ class CalendarDayEntity
         $this->day       = $day;
         $this->isCurrent = $isCurrent;
         $this->timestamp = $timestamp;
-    }
-
-    /**
-     * @param string  $key
-     *
-     * @return mixed
-     */
-    public function __get(string $key)
-    {
-        return $this->{$key};
-    }
-
-    /**
-     * @return array
-     */
-    public function toArray()
-    {
-        $reflect = new \ReflectionClass($this);
-        $properties = $reflect->getProperties(\ReflectionProperty::IS_PROTECTED);
-
-        $array = [];
-
-        foreach ($properties as $property) {
-            $property->setAccessible(true);
-            $array[$property->getName()] = $property->getValue($this);
-        }
-
-        return $array;
     }
 }
